authorization

+++ app/Http/Controllers/TasklistsController.php
@@ -16,12 +16,22 @@ class TasklistsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        $tasklists = Tasklist::all();
+    {   
+        $tasklist= [];
+         if (\Auth::check()) {
+             
+            $user=\Auth::user();
+            $tasklists  = $user->tasklist()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $tasklist = [
+                'user'=>$user,
+                'tasklists'=>$tasklists,
+                ];
+            return view('tasklists.index',$tasklist);
+        }else {
+            return view('welcome');
+        }
 
-        return view('tasklists.index', [
-            'tasklists' => $tasklists,
-        ]);
     }
 
     /**
@@ -31,11 +41,14 @@ class TasklistsController extends Controller
      */
     public function create()
     {
+        if (\Auth::check()) {
         $tasklist = new Tasklist;
 
         return view('tasklists.create', [
             'tasklist' => $tasklist,
         ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -54,6 +67,7 @@ class TasklistsController extends Controller
         $tasklist = new Tasklist;
         $tasklist->status = $request-> status;
         $tasklist->content = $request->content;
+        $tasklist->user_id = \Auth::user()->id;
         $tasklist->save();
 
         return redirect('/');
@@ -81,12 +95,22 @@ class TasklistsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
+    
     {
-        $tasklist = Tasklist::find($id);
-
-        return view('tasklists.edit', [
+     if (!empty($tasklist->user_id)) {
+            if (\Auth::user()->id === $tasklist->user_id) {
+           return view('tasklists.edit', [
             'tasklist' => $tasklist,
-        ]);
+            ]);
+             }
+            else {
+                return redirect('/');
+            }
+        }
+        
+        else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -121,7 +145,11 @@ class TasklistsController extends Controller
     public function destroy($id)
     {
         $tasklist = Tasklist::find($id);
+        
+        if (\Auth::user()->id === $tasklist->user_id) {
         $tasklist->delete();
+            
+        }
 
         return redirect('/');
     }
